@@ -13,28 +13,29 @@
     "deploy": "ng deploy"
   },
   "engines": {
-    "node": "^18.13.0 || ^20.9.0",
-    "npm": ">=8.0.0"
+    "node": "^18.19.0 || ^20.9.0",
+    "npm": ">=8.0.0",
+    "typescript": "^5.4.0"
   },
   "dependencies": {
-    "@angular/animations": "^17.2.4",
+    "@angular/animations": "^18.2.0",
     "@angular/cdk": "^17.2.2",
-    "@angular/common": "^17.2.4",
-    "@angular/compiler": "^17.2.4",
-    "@angular/core": "^17.2.4",
-    "@angular/forms": "^17.2.4",
+    "@angular/common": "^18.2.0",
+    "@angular/compiler": "^18.2.0",
+    "@angular/core": "^18.2.0",
+    "@angular/forms": "^18.2.0",
     "@angular/material": "^17.2.2",
-    "@angular/platform-browser": "^17.2.4",
-    "@angular/platform-browser-dynamic": "^17.2.4",
-    "@angular/router": "^17.2.4",
+    "@angular/platform-browser": "^18.2.0",
+    "@angular/platform-browser-dynamic": "^18.2.0",
+    "@angular/router": "^18.2.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.5.0",
-    "zone.js": "~0.14.4"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.2.3",
-    "@angular/cli": "^17.2.3",
-    "@angular/compiler-cli": "^17.2.4",
+    "@angular-devkit/build-angular": "^18.2.0",
+    "@angular/cli": "^18.2.0",
+    "@angular/compiler-cli": "^18.2.0",
     "@types/jasmine": "~4.3.1",
     "angular-cli-ghpages": "^1.0.5",
     "jasmine-core": "~4.5.0",
@@ -44,6 +45,6 @@
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
     "tailwindcss": "^3.2.4",
-    "typescript": "~5.3.3"
+    "typescript": "~5.5.4"
   }
 }
